add authorize service with rabbitmq

+++ backend/src/main/java/com/example/backend/controller/mysql/FoodController.java
@@ -4,13 +4,14 @@ import com.example.backend.controller.mysql.body.FoodBody;
 import com.example.backend.controller.mysql.response.FoodResponse;
 import com.example.backend.controller.mysql.response.Options;
 import com.example.backend.pojo.mysql.Food;
+import com.example.backend.requestBody.AuthTokenData;
 import com.example.backend.service.mysql.FoodService;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.swing.text.html.Option;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -21,6 +22,9 @@ public class FoodController {
 
     @Autowired
     private FoodService foodService;
+
+    @Autowired
+    private RabbitTemplate rabbitTemplate;
 
     @GetMapping("/getFood")
     public ResponseEntity allFood() {
@@ -55,29 +59,41 @@ public class FoodController {
     }
 
     @PatchMapping("/addFood")
-    public ResponseEntity<?> addFood(@RequestBody FoodBody foodBody) {
-        if (foodService.addFood(foodBody)) {
-            return new ResponseEntity<>("Add food successfully.", HttpStatus.OK);
-        } else {
+    public ResponseEntity<?> addFood(@RequestBody FoodBody foodBody, @RequestHeader("token") String token, @RequestHeader("userId") String userId) {
+        AuthTokenData header = new AuthTokenData(token, userId);
+        Object auth = rabbitTemplate.convertSendAndReceive("AuthExchange", "auth", header);
+        if ((boolean) auth) {
+            if (foodService.addFood(foodBody)) {
+                return new ResponseEntity<>("Add food successfully.", HttpStatus.OK);
+            }
             return new ResponseEntity<>("Something went wrong.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PatchMapping("/updFood/{id}")
-    public ResponseEntity<?> updateFoodDataById(@PathVariable("id") int id, @RequestBody FoodBody foodBody) {
-        if (foodService.updateFoodDataById(id, foodBody)) {
-            return new ResponseEntity<>("Add accessory successfully.", HttpStatus.OK);
-        } else {
+    public ResponseEntity<?> updateFoodDataById(@PathVariable("id") int id, @RequestBody FoodBody foodBody, @RequestHeader("token") String token, @RequestHeader("userId") String userId) {
+        AuthTokenData header = new AuthTokenData(token, userId);
+        Object auth = rabbitTemplate.convertSendAndReceive("AuthExchange", "auth", header);
+        if ((boolean) auth) {
+            if (foodService.updateFoodDataById(id, foodBody)) {
+                return new ResponseEntity<>("Add accessory successfully.", HttpStatus.OK);
+            }
             return new ResponseEntity<>("Something went wrong.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/delFood/{id}")
-    public ResponseEntity<?> deleteByFoodId(@PathVariable("id") int id) {
-        if (foodService.deleteByFoodId(id)) {
-            return new ResponseEntity<>("Delete food successfully", HttpStatus.OK);
-        } else {
+    public ResponseEntity<?> deleteByFoodId(@PathVariable("id") int id, @RequestBody FoodBody foodBody, @RequestHeader("token") String token, @RequestHeader("userId") String userId) {
+        AuthTokenData header = new AuthTokenData(token, userId);
+        Object auth = rabbitTemplate.convertSendAndReceive("AuthExchange", "auth", header);
+        if ((boolean) auth) {
+            if (foodService.deleteByFoodId(id)) {
+                return new ResponseEntity<>("Delete food successfully", HttpStatus.OK);
+            }
             return new ResponseEntity<>("Something went wrong.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
